+++ src/components/ElementEditor/ElementEditor.tsx
@@ -1,7 +1,19 @@
 import React, { ChangeEvent } from 'react';
 import { SelectableValue } from '@grafana/data';
-import { Button, InlineField, InlineFieldRow, Input, RadioButtonGroup, Select } from '@grafana/ui';
 import {
+  Button,
+  CodeEditor,
+  Field,
+  InlineField,
+  InlineFieldRow,
+  Input,
+  RadioButtonGroup,
+  Select,
+  useStyles2,
+} from '@grafana/ui';
+import {
+  CodeEditorHeight,
+  CodeLanguage,
   CodeLanguageOptions,
   FormElementType,
   FormElementTypeOptions,
@@ -9,8 +21,9 @@ import {
   StringElementOptions,
   TestIds,
 } from '../../constants';
-import { FormElement } from '../../types';
+import { LocalFormElement } from '../../types';
 import { FormatNumberValue, GetElementWithNewType, ToNumberValue } from '../../utils';
+import { Styles } from './styles';
 
 /**
  * Properties
@@ -19,18 +32,18 @@ interface Props {
   /**
    * Element
    */
-  element: FormElement;
+  element: LocalFormElement;
 
   /**
    * On Change
    */
-  onChange: (element: FormElement, checkConflict?: boolean) => void;
+  onChange: (element: LocalFormElement, checkConflict?: boolean) => void;
 
   /**
    * On Change Option
    */
   onChangeOption: (
-    element: FormElement,
+    element: LocalFormElement,
     updatedOption: SelectableValue,
     value?: SelectableValue,
     checkConflict?: boolean
@@ -46,6 +59,11 @@ interface Props {
  * Element Editor
  */
 export const ElementEditor: React.FC<Props> = ({ element, onChange, onChangeOption, layoutSectionOptions }) => {
+  /**
+   * Styles
+   */
+  const styles = useStyles2(Styles);
+
   /**
    * Return
    */
@@ -339,7 +357,7 @@ export const ElementEditor: React.FC<Props> = ({ element, onChange, onChangeOpti
         element.type === FormElementType.SELECT ||
         element.type === FormElementType.MULTISELECT ||
         element.type === FormElementType.DISABLED) && (
-        <div>
+        <div className={styles.optionsContainer}>
           {element.options?.map((option, index) => (
             <InlineFieldRow key={index} data-testid={TestIds.formElementsEditor.fieldOption(option.value)}>
               <InlineField label="Type" labelWidth={8}>
@@ -444,6 +462,23 @@ export const ElementEditor: React.FC<Props> = ({ element, onChange, onChangeOpti
           </Button>
         </div>
       )}
+
+      <Field label="Show if returned value is true">
+        <CodeEditor
+          value={element.showIf || ''}
+          language={CodeLanguage.JAVASCRIPT}
+          height={`${CodeEditorHeight}px`}
+          onBlur={(code) => {
+            onChange({
+              ...element,
+              showIf: code,
+            });
+          }}
+          monacoOptions={{ formatOnPaste: true, formatOnType: true }}
+          showLineNumbers={true}
+          aria-label={TestIds.formElementsEditor.fieldShowIf}
+        />
+      </Field>
     </>
   );
 };