+++ src/types/request.ts
@@ -69,7 +69,7 @@ export interface RequestOptions {
   confirm: boolean;
 
   /**
-   * @deprecated Keep for backward compatability.
+   * @deprecated Keep for backward compatibility.
    * Add only Updated Elements to Payload
    *
    * @type {boolean}