+++ src/utils/request.ts
@@ -23,7 +23,7 @@ export const GetPayloadForRequest = ({
   }
 
   /**
-   * Use deprecated property for backward compatability
+   * Use deprecated property for backward compatibility
    */
   const updatedOnly =
     request.payloadMode === undefined ? request.updatedOnly : request.payloadMode === PayloadMode.UPDATED;