+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Update jest selectors to use npm package (#209)
 - Add onChange to update elements in local state within custom code (#214)
+- Update ESLint configuration (#215)
 
 ## 3.0.0 (2023-07-15)